Fixed a NPE
git-svn-id: https://svn.wso2.org/repos/wso2/trunk/tools@173559 a5903396-d722-0410-b921-86c7d4935375

+++ ide/eclipse/common/org.wso2.developerstudio.eclipse.utils/src/org/wso2/developerstudio/eclipse/utils/file/FileUtils.java
@@ -117,15 +117,18 @@ public class FileUtils{
 
 	public static void deleteDirectories(File dir) {
 		File[] children = dir.listFiles();
-		for (int i = 0; i < children.length; i++){
-			if (children[i].list() != null && children[i].list().length > 0){
-				deleteDirectories(children[i]);
-			} else if(children[i].isDirectory()){
-				deleteDirectory(children[i]);
-			}else  if(children[i].isFile()){
-				boolean isDeleted = org.apache.commons.io.FileUtils.deleteQuietly(children[i]);
-				if(!isDeleted){
-					children[i].deleteOnExit();
+		if (children != null) {
+			for (int i = 0; i < children.length; i++) {
+				if (children[i].list() != null && children[i].list().length > 0) {
+					deleteDirectories(children[i]);
+				} else if (children[i].isDirectory()) {
+					deleteDirectory(children[i]);
+				} else if (children[i].isFile()) {
+					boolean isDeleted = org.apache.commons.io.FileUtils
+							.deleteQuietly(children[i]);
+					if (!isDeleted) {
+						children[i].deleteOnExit();
+					}
 				}
 			}
 		}
